CMakeLists.txt: initialize the project first
Initialize the project before including all modules and do not
enable languages on their own.

The old behavior had the effect of not initializing the
CMAKE_INSTALL_PREFIX on Windows (and perhaps other variables?) to
the correct value.

Also, set CMake policy CMP0102 to NEW as the CMocka include
and library directory variables could be uninitialized between
subsequent runs of CMake from the GUI.